requirements.txt
python-telegram-bot==21.7 
google-auth==2.23.0
apscheduler==3.10.4
python-dotenv==1.0.0
Pillow==10.4.0  